+++ packages/collectors/src/index.ts
@@ -1,0 +1,7 @@
+import 'dotenv/config'
+import { startDiscoveryCron } from './polymarket/discovery'
+import { startStream } from './polymarket/stream'
+
+console.log('Starting collectors...')
+startDiscoveryCron()
+startStream()